rename Node and Node2
